fix: z-index 추가, image trasnsform시 header와 겹치는 문제 해결
