+++ src/components/CardListItem/CardListItem.module.css
@@ -1,0 +1,95 @@
+.wrapper {
+  /* max-width: 1540px; */
+  padding: 0px 200px 40px;
+  margin: 0 auto;
+}
+
+.l_row {
+  display: flex;
+  flex-wrap: wrap;
+  margin: 0 -10px;
+  row-gap: 20px;
+}
+
+.l_col {
+  box-sizing: border-box;
+  padding: 0 10px;
+  width: 33.33333%;
+}
+
+.link {
+  border-radius: 10px;
+  overflow: hidden;
+  box-shadow: 0 0 10px 3px rgba(0, 0, 0, 0.1);
+}
+
+.link:hover .link-cover img {
+  transform: scale(1.3);
+  z-index: 0;
+}
+
+.link-contents {
+  padding: 15px;
+}
+
+.link-cover {
+  background-color: var(--Linkbrary-bg);
+  aspect-ratio: 16/9;
+  overflow: hidden;
+}
+
+.link-cover img {
+  height: 100%;
+  width: 100%;
+  object-fit: cover;
+  transition: 0.1s linear;
+}
+
+/* .link-cover img:hover {
+    transform: scale(1.3);
+  } */
+
+.link-title {
+  font-size: 1.6rem;
+  margin-bottom: 10px;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.link-description {
+  font-size: 1.2rem;
+  line-height: 1.3;
+  height: calc(2 * 1.3em);
+  overflow: hidden;
+}
+
+.link-update {
+  margin-bottom: 10px;
+  font-size: 1.2rem;
+}
+
+.link-date {
+  margin-top: 10px;
+  font-size: 1.2rem;
+}
+
+@media (max-width: 1199px) {
+  .wrapper {
+    max-width: 832px;
+    padding: 0px 32px 40px;
+    margin: 0 auto;
+  }
+}
+
+@media (max-width: 1124px) {
+  .l_col {
+    width: 50%;
+  }
+}
+
+@media (max-width: 580px) {
+  .l_col {
+    width: 100%;
+  }
+}
