+++ src/components/Header/Header.module.css
@@ -6,6 +6,7 @@
   top: 0;
   height: 9em;
   background-color: var(--Linkbrary-bg);
+  z-index: 1;
 }
 
 .headings {
